+++ CHANGELOG.md
@@ -16,11 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## Unreleased
-Add 'FLAG' attribute to Episciences::Paper class
-Alter table PAPERS: add new field 'TAG
-New script to update papers table
+- Add 'FLAG' attribute to Episciences::Paper class
+- Alter table PAPERS: add new field 'TAG
+- New script to update papers table
+- Update paper stats only if paper is published and user is not the contributor
 
 ## Unreleased
+- Update paper stats only if paper is published and user is not the contributor
+
+## 1.0.16 - 2021-09-13
 ### Fixed
 - Fixed issue: [64](https://github.com/CCSDForge/episciences/issues/64): Layout of mails with reviewers' comments
 - Improvement of the process of accounts merging